requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.5.3
Django==4.2.4
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.6
idna==3.7
inflection==0.5.1
packaging==24.0
psycopg2==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
